requirements.txt
requests>=2.31.0
pillow>=10.0.0
streamlit>=1.29.0
